modules/locationsws/app_consumer/requirements-consumer.txt
bson==0.5.10
kafka-python==2.0.2
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==2.0.10
psycopg2==2.9.6
GeoAlchemy2==0.13.2
shapely==2.0.1
